Revert "fix: move back storefrontapp occ config to 'app.module.ts', and restore b2b- and b2c-configuration.module.ts (#21120)"
This reverts commit 70bc432142621a74b3509c1c7900b323fc658095.
