+++ projects/storefrontapp/src/app/app.module.ts
@@ -8,9 +8,7 @@ import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { EffectsModule } from '@ngrx/effects';
 import { StoreModule } from '@ngrx/store';
-import { provideConfig } from '@spartacus/core';
 import { AppRoutingModule } from '@spartacus/storefront';
-import { environment } from '../environments/environment';
 import { privateProviders } from './private/private.providers';
 import { SpartacusModule } from './spartacus/spartacus.module';
 
@@ -22,18 +20,6 @@ import { SpartacusModule } from './spartacus/spartacus.module';
     EffectsModule.forRoot([]),
     SpartacusModule,
   ],
-  providers: [
-    provideConfig({
-      backend: {
-        occ: {
-          // Note: The next line is edited by our internal script in the `ec-automate-pipelines` repo. Don't move it to other file.
-          baseUrl: environment.occBaseUrl,
-          prefix: environment.occApiPrefix,
-        },
-      },
-    }),
-
-    privateProviders,
-  ],
+  providers: [privateProviders],
 })
 export class AppModule {}
