+++ projects/storefrontapp/src/app/private/private.providers.ts
@@ -18,6 +18,14 @@ import { TestOutletModule } from '../../test-outlets/test-outlet.module';
  */
 export const privateProviders = makeEnvironmentProviders([
   provideConfig({
+    // Custom OCC backend configuration:
+    backend: {
+      occ: {
+        baseUrl: environment.occBaseUrl,
+        prefix: environment.occApiPrefix,
+      },
+    },
+
     // Custom routing configuration for e2e testing:
     routing: {
       routes: {
